requirements.txt
python-pptx==0.6.17
pandas==0.24.2